Base Model and dto class added

+++ parkinglot/src/main/java/models/BaseModel.java
@@ -3,10 +3,8 @@ package models;
 import java.time.LocalDateTime;
 
 public class BaseModel {
-    private  int id;
+    private  long id;
     private LocalDateTime CreatedAt;
     private LocalDateTime UpdatedAt;
-
-
 
 }
